06_trigger_word/requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
sounddevice>=0.4.5
webrtcvad>=2.0.10
numpy>=1.21.0
scipy>=1.7.0
# Windows users only: pip install pyttsx3
